Refactored testing and mocking, and introduced full membership module mock
